@@ -1,22 +1,28 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Remove="Serializers\**" />
+    <Compile Remove="Servers\**" />
+    <EmbeddedResource Remove="Serializers\**" />
+    <EmbeddedResource Remove="Servers\**" />
+    <None Remove="Serializers\**" />
+    <None Remove="Servers\**" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="MediatR" Version="12.1.1" />
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="7.0.1" />
+    <PackageReference Include="Microsoft.Extensions.Hosting" Version="8.0.0-rc.1.23419.4" />
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Townsharp.Core\Townsharp.Core.csproj" />
     <ProjectReference Include="..\Townsharp.Infrastructure\Townsharp.Infrastructure.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Servers\" />
   </ItemGroup>
 
 </Project>
