@@ -1,25 +1,27 @@
 ﻿using System.Text.Json.Nodes;
 
 using Microsoft.Extensions.Logging;
+
 using Townsharp.Consoles.Commands;
-using Townsharp.Infrastructure.GameConsole;
+using Townsharp.Infrastructure.GameConsoles;
 using Townsharp.Infrastructure.ServerConsole;
 using Townsharp.Internals.Consoles;
+using Townsharp.Servers;
 
 namespace Townsharp.Consoles;
 
 public class GameServerConsole
 {
-    private readonly GameServerId id;
+    private readonly ServerId id;
     private readonly ConsoleClientFactory consoleClientFactory;
     private readonly ConsoleAccessProvider consoleAccessProvider;
     private readonly ILogger<GameServerConsole> logger;
     private Task<ConsoleClient?> consoleClientFactoryTask = Task.FromResult<ConsoleClient?>(null);
 
-    private readonly SemaphoreSlim semaphore = new SemaphoreSlim(1, 1);
+    private readonly SemaphoreSlim semaphore = new(1, 1);
 
     internal GameServerConsole(
-        GameServerId id,
+        ServerId id,
         ConsoleClientFactory consoleClientFactory,
         ConsoleAccessProvider consoleAccessProvider,
         ILogger<GameServerConsole> logger)
