@@ -1,10 +1,11 @@
-﻿using System.Net.Http.Headers;
+﻿using System.Threading.Channels;
 
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
-using Townsharp.Infrastructure.GameConsole;
-using Townsharp.Infrastructure.ServerConsole;
+using Townsharp.Infrastructure.Consoles;
+using Townsharp.Infrastructure.Consoles.Models;
+using Townsharp.Infrastructure.GameConsoles;
 using Townsharp.Infrastructure.WebApi;
 
 public class ConsoleRepl : IHostedService
@@ -35,9 +36,10 @@ public class ConsoleRepl : IHostedService
             bool retryNeeded = false;
             UriBuilder uriBuilder = new UriBuilder();
             string accessToken = "";
+            Channel<ConsoleEvent> eventChannel = Channel.CreateUnbounded<ConsoleEvent>();
             try
             {
-                var response = await this.webApiClient.RequestConsoleAccessAsync(ulong.Parse(serverId!));
+                var response = await this.webApiClient.RequestConsoleAccessAsync(int.Parse(serverId!));
 
                 if (!response["allowed"]?.GetValue<bool>() ?? false)
                 {
@@ -66,25 +68,20 @@ public class ConsoleRepl : IHostedService
                         
             try
             {
-                ConsoleClient consoleClient = await this.consoleClientFactory.CreateAndConnectAsync(uriBuilder.Uri, accessToken);
+                ConsoleClient consoleClient = this.consoleClientFactory.CreateClient(uriBuilder.Uri, accessToken, eventChannel.Writer);
+                await consoleClient.ConnectAsync(cancellationTokenSource.Token);
                 Task getCommandsTask = this.GetCommandsAsync(consoleClient, cancellationTokenSource.Token); // this doesn't work right, stupid sync console.
 
-                void handleGameConsoleEvent(object? sender, GameConsoleEvent e)
+                await foreach (var consoleEvent in eventChannel.Reader.ReadAllAsync(cancellationTokenSource.Token))
                 {
-                    this.logger.LogInformation(e.ToString());
-                }
+                    var message = consoleEvent switch
+                    {
+                        ConsoleEvent e => e.ToString(),
+                        _ => "Not Implemented"
+                    };
 
-                void handleDisconnected (object? sender,  EventArgs e)
-                {
-                    cancellationTokenSource.Cancel();
-                    this.logger.LogInformation("Disconnected from server {serverId}. Will attempt to reconnect.", serverId);
-                    consoleClient.Dispose();
-                    consoleClient.GameConsoleEventReceived -= handleGameConsoleEvent;
-                    consoleClient.Disconnected -= handleDisconnected;
+                    Console.WriteLine(message);
                 }
-
-                consoleClient.GameConsoleEventReceived += handleGameConsoleEvent;
-                consoleClient.Disconnected += handleDisconnected;
 
                 await getCommandsTask;
             }
@@ -109,7 +106,14 @@ public class ConsoleRepl : IHostedService
 
             var result = await consoleClient.RunCommand(command!, TimeSpan.FromSeconds(30), token);
 
-            Console.WriteLine(result.ToString());
+            if (result.IsCompleted)
+            {
+                Console.WriteLine(result?.Message?.data?.ToString());
+            }
+            else
+            {
+                Console.WriteLine(result.ToString());
+            }
         }
     }
 
