@@ -1,6 +1,6 @@
-﻿namespace Townsharp.Infrastructure.Identity.Models;
+﻿namespace Townsharp.Infrastructure.Configuration;
 
 public record UserCredential(string Username, string PasswordHash)
 {
-    public bool IsConfigured => !(String.IsNullOrEmpty(this.Username) || String.IsNullOrEmpty(this.PasswordHash));
+    public bool IsConfigured => !(string.IsNullOrEmpty(this.Username) || string.IsNullOrEmpty(this.PasswordHash));
 }
