@@ -1,9 +1,8 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 
 using Townsharp.Infrastructure.Configuration;
-using Townsharp.Infrastructure.GameConsole;
+using Townsharp.Infrastructure.GameConsoles;
 using Townsharp.Infrastructure.Identity;
-using Townsharp.Infrastructure.Identity.Models;
 using Townsharp.Infrastructure.Subscriptions;
 using Townsharp.Infrastructure.WebApi;
 
@@ -11,6 +10,7 @@ namespace Townsharp.Infrastructure.Hosting;
 
 public static class ServiceCollectionExtensions
 {
+    // This needs to die off, either we expose standalone clients via a dedicated library, via this library, or we just support it via townsharp.
     public static IServiceCollection AddTownsharpInfra(this IServiceCollection services)
     {
         services.AddHttpClient();
