@@ -4,7 +4,7 @@ public interface IBotTokenProvider
 {
     public ValueTask<string> GetTokenAsync(CancellationToken cancellationToken = default);
 
-    public ValueTask<ulong> GetBotUserIdAsync(CancellationToken cancellationToken = default);
+    public ValueTask<int> GetBotUserIdAsync(CancellationToken cancellationToken = default);
 
     public ValueTask<string> GetBotUserNameAsync(CancellationToken cancellationToken = default);
 
