@@ -3,7 +3,7 @@
 using Polly;
 using Polly.Retry;
 
-using Townsharp.Infrastructure.Identity.Models;
+using Townsharp.Infrastructure.Configuration;
 using Townsharp.Infrastructure.Utilities;
 
 namespace Townsharp.Infrastructure.Identity;
