@@ -6,13 +6,14 @@ using MediatR;
 using Microsoft.Extensions.Logging;
 
 using Townsharp.Consoles;
-using Townsharp.Infrastructure.GameConsole;
+using Townsharp.Infrastructure.GameConsoles;
+using Townsharp.Servers;
 
 namespace Townsharp.Internals.Consoles;
 
-internal class GameServerConsoleManager : IReadOnlyDictionary<GameServerId, GameServerConsole>
+internal class GameServerConsoleManager : IReadOnlyDictionary<ServerId, GameServerConsole>
 {
-    private readonly Dictionary<GameServerId, GameServerConsole> gameServerConsoles;
+    private readonly Dictionary<ServerId, GameServerConsole> gameServerConsoles;
     private readonly IMediator mediator;
     private readonly ConsoleClientFactory consoleClientFactory;
     private readonly ConsoleAccessProvider consoleAccessProvider;
@@ -24,28 +25,28 @@ internal class GameServerConsoleManager : IReadOnlyDictionary<GameServerId, Game
         ConsoleAccessProvider consoleAccessProvider, 
         ILoggerFactory loggerFactory)
     {
-        gameServerConsoles = new Dictionary<GameServerId, GameServerConsole>();
+        gameServerConsoles = new Dictionary<ServerId, GameServerConsole>();
         this.mediator = mediator;
         this.consoleClientFactory = consoleClientFactory;
         this.consoleAccessProvider = consoleAccessProvider;
         this.loggerFactory = loggerFactory;
     }
 
-    public GameServerConsole this[GameServerId key] => gameServerConsoles[key];
+    public GameServerConsole this[ServerId key] => gameServerConsoles[key];
 
-    public IEnumerable<GameServerId> Keys => gameServerConsoles.Keys;
+    public IEnumerable<ServerId> Keys => gameServerConsoles.Keys;
 
     public IEnumerable<GameServerConsole> Values => gameServerConsoles.Values;
 
     public int Count => gameServerConsoles.Count();
 
-    public bool ContainsKey(GameServerId key) => gameServerConsoles.ContainsKey(key);
+    public bool ContainsKey(ServerId key) => gameServerConsoles.ContainsKey(key);
 
-    public IEnumerator<KeyValuePair<GameServerId, GameServerConsole>> GetEnumerator() => gameServerConsoles.GetEnumerator();
+    public IEnumerator<KeyValuePair<ServerId, GameServerConsole>> GetEnumerator() => gameServerConsoles.GetEnumerator();
 
-    public bool TryGetValue(GameServerId key, [MaybeNullWhen(false)] out GameServerConsole value) => gameServerConsoles.TryGetValue(key, out value);
+    public bool TryGetValue(ServerId key, [MaybeNullWhen(false)] out GameServerConsole value) => gameServerConsoles.TryGetValue(key, out value);
 
-    internal Task ManageGameServerConsoleAsync(GameServerId serverId)
+    internal Task ManageGameServerConsoleAsync(ServerId serverId)
     {
         var console = new GameServerConsole(serverId, this.consoleClientFactory, this.consoleAccessProvider, this.loggerFactory.CreateLogger<GameServerConsole>());
         this.gameServerConsoles.Add(serverId, console);
