@@ -1,13 +1,14 @@
 ﻿using MediatR;
+using Townsharp.Servers;
 
 namespace Townsharp.Internals.Notifications;
 
 internal class GameServerManagedNotification : INotification
 {
-    public GameServerManagedNotification(GameServerId serverId)
+    public GameServerManagedNotification(ServerId serverId)
     {
         this.ServerId = serverId;
     }
 
-    public GameServerId ServerId { get; }
+    public ServerId ServerId { get; }
 }
